requirements.txt
fastapi
uvicorn[standard]
langchain-google-genai
elevenlabs
pydub
python-dotenv
langchain
pydantic-settings